+++ CHANGELOG.md
@@ -9,8 +9,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added 
 
-- Added a dataset reader for SQuAD 2.0, the `Squad2Reader` (registered as `squad2`).
+- Added dataset readers for SQuAD 2.0, the `Squad2Reader` (registered as `"squad2"`) and `TransformerSquad2Reader`
+  (registered as `"transformer_squad2"`).
 - Updated the SQuAD v1.1 metric to work with SQuAD 2.0 as well.
+- Updated the `TransformerQA` model to work for SQuAD 2.0 with the `TransformerSquad2Reader`.
 - Added official support for Python 3.8.
 - Added a json template for model cards.